PlanterAI-main/SensorApi/requirements.txt
bidict==0.21.2
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==1.16.0
eventlet==0.31.0
firebase-admin==5.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-SocketIO==5.0.3
gcloud==0.17.0
google-api-core==1.30.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.38.1
httplib2==0.19.1
idna==2.10
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
jws==0.1.3
MarkupSafe==2.0.1
msgpack==1.0.2
oauth2client==3.0.0
packaging==21.0
proto-plus==1.19.0
protobuf==4.0.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.6.6
pyFirmata==1.1.0
pyparsing==2.4.7
Pyrebase==3.0.27
pyserial==3.5
python-engineio==4.2.0
python-firebase==1.2.1
python-jwt==2.0.1
python-socketio==5.3.0
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
uWSGI==2.0.19.1
Werkzeug==2.0.1
zipp==3.4.1
